+++ bin/firefox-browser.sh
@@ -1,4 +1,0 @@
-#!/bin/sh
-echo "[INFO] Starting wrapped firefox browser"
-/usr/bin/firefox-original -headless $@
-