+++ bin/system-status.sh
@@ -1,11 +1,16 @@
 #!/bin/bash
 
-echo "[INFO] Docker RobotFramework Status"
-
+START_DATE=$(date)
+echo "[INFO] Docker RobotFramework Status start ${START_DATE}"
 
 echo "[INFO] Status of Python packages:"
-pip list --outdated  --format=columns
-
+PYTHON="$(pip list --outdated  --format=columns)"
+echo $PYTHON
+echo "$PYTHON">>$REPORTDIR/python-update.log
 echo "[INFO] Status of installed system packages"
-yum check-update
+SYSTEM="$(yum check-update)"
+echo "$SYSTEM"
+echo "$SYSTEM" >> $REPORTDIR/system-update.log
 
+END_DATE=$(date)
+echo "[INFO] Docker RobotFramework Status end ${END_DATE}"