+++ CHANGELOG.md
@@ -1,3 +1,10 @@
 # Extended Robot Framework in Docker and Chrome Headless 
  
+* 11.0.0
+	* extension of logging
+	* creation of index.html for test and documentation overview
+	* automatic library documentation
+	* automatic test documentation
+	* automatic test resource documentation
+	
 * 10.0.0 Initial Klone more libraries, reduced to chrome headless